rename index to webpack
