+++ .gitignore
@@ -17,3 +17,4 @@ node_modules/
 # Build Directory
 dist
 .DS_STORE
+.vscode
