+++ test/fixtures/webpack/main.tsx
@@ -1,6 +1,5 @@
 import React, {Component} from 'react';
 const img =require('file-loader!./1.png')
-// console.log(add(sub(1,2),1))
 
 
 export default class MainComponent extends Component {
